Score_entry: Correct touch/sprawl penalty amounts
Update 21 clarifies the penalty cost:
21 – CALCULATING ROOF PENALTIES
You can get up to eight sprawl and/or touch penalties, as represented by the roof debris on the mat.  The Missions page is unclear on the following fact:  Here is the cost for penalty accumulation:  1=10, 2=20, 3=30, 4=40, 5=53, 6=66, 7=79, 8=92.  Sorry for the confusion.  This far into the season I would normally be unable to make such a serious correction, but I trust no one was developing a penalty-based strategy.
Signed-off-by:Courtney Goeltzenleuchter <courtney@coloradofirst.org>