score_entry: Fix issue submitting unrecognized data
The values used by the Javascript score computation and the
server based score computation are slightly different.
In particular, there should be no input values with a
space in them. The original score entry simply set the label
and value to the same thing, i.e. Lt_Blue. With the addition
of the javascript computation the form could now have a
different label and value, i.e. the label is "Lt Blue" but the value
sent to the server should be "Lt_Blue". However, that wasn't
working quite right. This commit fixes that issue.