Version 1.2
BUGFIX:  Occasionally the pre-match audio sound effect would play when
there was zero seconds set for the pre-match timer.  This has been
corrected.

BUGFIX:  Display would slow down its update rate as more and more scores
are entered, and the last page of the scores (lowest teams) would
display much longer than the first page (highest teams).  Now processing
the score information in parallel to displaying the scores.  This may
mean scores will have to cycle twice before being updated.

FEATURE:  Display has the "default" ability to translate negative scores
to zero, so that negative scores are not seen (and so instead a score of
zero is displayed).  This is an option in the display configuration
window.

FEATURE:  Display has the ability to play game sounds, allowing you to
play the sounds in a remote location away from the main scoring
computer.  Sounds are sourced from the remote computer, so any
additional/replacement sounds must be added to all machines playing the
sounds.

FEATURE:  New sound effect has been added, a "hurry up" sound effect
that plays when there is 30 seconds remaining in the match.  The "hurry
up" sound effect is the "TimeToRun" sound effect that sounds like an
animated character running quickly in place (as if it has no traction).
Also known as the "Scooby Doo Running" sound effect.

FEATURE:  Sound configuration and timer configuration have been moved to
a button-menu within the Profile window.  The button window is an
auto-save window, and will persist values between software launches.