ScoreEntry: Re-use form to speed processing
On tablet devices the initial build of the form is pretty slow.
With this change the form is reset after score submission to
avoid the start-up cost.
Adds Reset Form button to clear all selections (also used to reset
the form after submission).
Adds support for manual score entry mode as well.