ScoreEntry: Use cached score_info
When putting together the HTML for the questions the code also
saves an array of the critical information. This array is now
used for all the functions that need to scan through questions
such as validate, defaults, reset, compute score, etc. Hopefully
this is faster than using the XML methods to extract the data
from the Data.xml object.