Version 1.3
FEATURE:  Added simple security to Score Entry and Timekeeper functions.
Remote systems requesting access to these pages/functions will be
required to provide a password the first time they attempt to access the
pages - this password is set from the SETTINGS page in the software.
The local machine never has to enter a password, however.

FEATURE:  Score Entry webpage has been optimized.  It is faster on load,
and each score submission doesn't require the page to reload any more.
You will still receive a notification when the score is entered upon
submit, however.

FEATURE:  Section numbers have been added to the header titles for each
item, to help make the score entry more like the score sheets.

FEATURE:  Added "Save on Exit" feature that asks if you want to create a
zip file backup when you attempt to exit the software.  Tournaments
should use this to save their data to a USB Flash Drive once they
complete the day's tournament.

FEATURE:  Significantly lowered the system resources required by the
Server and Display.

FEATURE:  Added the ability to reset the score entry web page using a
button at the top of the page.

CHANGE:  The displays only load the logos from the server once ever,
when the display is connected to the server.  If you change any of the
logos for the display once the display is connected, you must restart
each display you wish to have the changed logos on.

CHANGE:  The Score Entry webpage no longer requires a reload after a
score is entered.  Added a pop-up alert to let you know once the score
had been sent to the server.

CHANGE:  Altered titles of some web pages.